Remove Deprecated Features
Move to a model where we build whatever branches are checked out
in the source directories.  This was a bit too smart (and complicated)
in handling details best left to the user and the revision control
system.

In addition, get rid of support for llvm-gcc and building gcc as
these are no longer necessary.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk/utils@149149 91177308-0d34-0410-b5e6-96231b3b80d8