Add CodeGenTarget::guessInstructionProperties.
Currently, TableGen just guesses instruction properties when it can't
infer them form patterns.

This adds a guessInstructionProperties flag to the instruction set
definition that will be used to disable guessing. The flag is intended
as a migration aid. It will be removed again when no more targets need
their properties guessed.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk/utils@162460 91177308-0d34-0410-b5e6-96231b3b80d8