Add a completely hack-ish tool to sort includes according to the coding standards.
I am a terrible Python programmer. Patches more the welcome. Please tell
me how this should look if it should look differently. It's just a tiny
little script so it didn't make sense to go through pre-commit review,
especially as someone who actually knows python may want to just rip it
apart and do it The Right Way.

I will be preparing a commit shortly that uses this script to
canonicalize *all* of the #include lines in LLVM. Really, all of them.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk/utils@169125 91177308-0d34-0410-b5e6-96231b3b80d8